+++ src/Component/updateForm/Form.css
@@ -1,0 +1,37 @@
+span {
+  background-color: #fff;
+  padding: 5px 10px;
+  border: 2px solid rgba(169, 165, 165, 0.813);
+  border-radius: 2px;
+  color: black;
+}
+
+.update {
+  position: absolute;
+  top: 0;
+  animation: slideInUp 1s ease-in-out 1;
+  transform: translateY(50%);
+}
+
+.Bg {
+  z-index: 1;
+  position: absolute;
+  width: 100vw;
+  height: 100vh;
+  background-color: #bcb9b98d;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  top: 0;
+}
+
+@keyframes slideInUp {
+  from {
+    transform: translate3d(0, 0, 0);
+    visibility: visible;
+  }
+
+  to {
+    transform: translate3d(0, 50%, 0);
+  }
+}
